poc/strands-agents/pyproject.toml
[project]
name = "poc-strands-agents"
version = "0.1.0"
description = "AWS Bedrock AgentCore (Strands Agents) PoC"
requires-python = ">=3.11"

dependencies = [
    # Strands Agents (AgentCore Python SDK)
    "strands-agents>=1.20.0",
    
    # Backend (workspace内依存)
    "agentcore-poc-backend",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["strands_poc"]

[tool.uv.sources]
agentcore-poc-backend = { workspace = true }
